refactor: remove pkg/extensions/search/common and move the code to the appropriate packages
