+++ pkg/extensions/search/search_test.go
@@ -1,7 +1,7 @@
 //go:build search
 // +build search
 
-package common_test
+package search_test
 
 import (
 	"context"
@@ -31,9 +31,10 @@ import (
 	"zotregistry.io/zot/pkg/api"
 	"zotregistry.io/zot/pkg/api/config"
 	"zotregistry.io/zot/pkg/api/constants"
+	"zotregistry.io/zot/pkg/common"
 	extconf "zotregistry.io/zot/pkg/extensions/config"
 	"zotregistry.io/zot/pkg/extensions/monitoring"
-	"zotregistry.io/zot/pkg/extensions/search/common"
+	"zotregistry.io/zot/pkg/extensions/search/convert"
 	cveinfo "zotregistry.io/zot/pkg/extensions/search/cve"
 	cvemodel "zotregistry.io/zot/pkg/extensions/search/cve/model"
 	"zotregistry.io/zot/pkg/log"
@@ -42,6 +43,7 @@ import (
 	"zotregistry.io/zot/pkg/storage/local"
 	. "zotregistry.io/zot/pkg/test"
 	"zotregistry.io/zot/pkg/test/mocks"
+	ocilayout "zotregistry.io/zot/pkg/test/oci-layout"
 )
 
 const (
@@ -146,36 +148,36 @@ type ImageSummaryResult struct {
 	Errors             []ErrorGQL         `json:"errors"`
 }
 
-func getTags() ([]common.TagInfo, []common.TagInfo) {
-	tags := make([]common.TagInfo, 0)
+func getTags() ([]cvemodel.TagInfo, []cvemodel.TagInfo) {
+	tags := make([]cvemodel.TagInfo, 0)
 
-	firstTag := common.TagInfo{
+	firstTag := cvemodel.TagInfo{
 		Name: "1.0.0",
-		Descriptor: common.Descriptor{
+		Descriptor: cvemodel.Descriptor{
 			Digest:    "sha256:eca04f027f414362596f2632746d8a178362170b9ac9af772011fedcc3877ebb",
 			MediaType: ispec.MediaTypeImageManifest,
 		},
 		Timestamp: time.Now(),
 	}
-	secondTag := common.TagInfo{
+	secondTag := cvemodel.TagInfo{
 		Name: "1.0.1",
-		Descriptor: common.Descriptor{
+		Descriptor: cvemodel.Descriptor{
 			Digest:    "sha256:eca04f027f414362596f2632746d8a179362170b9ac9af772011fedcc3877ebb",
 			MediaType: ispec.MediaTypeImageManifest,
 		},
 		Timestamp: time.Now(),
 	}
-	thirdTag := common.TagInfo{
+	thirdTag := cvemodel.TagInfo{
 		Name: "1.0.2",
-		Descriptor: common.Descriptor{
+		Descriptor: cvemodel.Descriptor{
 			Digest:    "sha256:eca04f027f414362596f2632746d8a170362170b9ac9af772011fedcc3877ebb",
 			MediaType: ispec.MediaTypeImageManifest,
 		},
 		Timestamp: time.Now(),
 	}
-	fourthTag := common.TagInfo{
+	fourthTag := cvemodel.TagInfo{
 		Name: "1.0.3",
-		Descriptor: common.Descriptor{
+		Descriptor: cvemodel.Descriptor{
 			Digest:    "sha256:eca04f027f414362596f2632746d8a171362170b9ac9af772011fedcc3877ebb",
 			MediaType: ispec.MediaTypeImageManifest,
 		},
@@ -184,7 +186,7 @@ func getTags() ([]common.TagInfo, []common.TagInfo) {
 
 	tags = append(tags, firstTag, secondTag, thirdTag, fourthTag)
 
-	vulnerableTags := make([]common.TagInfo, 0)
+	vulnerableTags := make([]cvemodel.TagInfo, 0)
 	vulnerableTags = append(vulnerableTags, secondTag)
 
 	return tags, vulnerableTags
@@ -1798,100 +1800,79 @@ func TestExpandedRepoInfo(t *testing.T) {
 
 func TestUtilsMethod(t *testing.T) {
 	Convey("Test utils", t, func() {
-		// Test GetRepo method
-		repo := common.GetRepo("test")
-		So(repo, ShouldEqual, "test")
-
-		repo = common.GetRepo(":")
-		So(repo, ShouldEqual, "")
-
-		repo = common.GetRepo("")
-		So(repo, ShouldEqual, "")
-
-		repo = common.GetRepo("test:123")
-		So(repo, ShouldEqual, "test")
-
-		repo = common.GetRepo("a/test:123")
-		So(repo, ShouldEqual, "a/test")
-
-		repo = common.GetRepo("a/test:123:456")
-		So(repo, ShouldEqual, "a/test")
-
 		// Test various labels
 		labels := make(map[string]string)
 
-		desc := common.GetDescription(labels)
+		desc := convert.GetDescription(labels)
 		So(desc, ShouldEqual, "")
 
-		license := common.GetLicenses(labels)
+		license := convert.GetLicenses(labels)
 		So(license, ShouldEqual, "")
 
-		vendor := common.GetVendor(labels)
+		vendor := convert.GetVendor(labels)
 		So(vendor, ShouldEqual, "")
 
-		categories := common.GetCategories(labels)
+		categories := convert.GetCategories(labels)
 		So(categories, ShouldEqual, "")
 
 		labels[ispec.AnnotationVendor] = "zot"
 		labels[ispec.AnnotationDescription] = "zot-desc"
 		labels[ispec.AnnotationLicenses] = "zot-license"
-		labels[common.AnnotationLabels] = "zot-labels"
+		labels[convert.AnnotationLabels] = "zot-labels"
 
-		desc = common.GetDescription(labels)
+		desc = convert.GetDescription(labels)
 		So(desc, ShouldEqual, "zot-desc")
 
-		license = common.GetLicenses(labels)
+		license = convert.GetLicenses(labels)
 		So(license, ShouldEqual, "zot-license")
 
-		vendor = common.GetVendor(labels)
+		vendor = convert.GetVendor(labels)
 		So(vendor, ShouldEqual, "zot")
 
-		categories = common.GetCategories(labels)
+		categories = convert.GetCategories(labels)
 		So(categories, ShouldEqual, "zot-labels")
 
 		labels = make(map[string]string)
 
 		// Use diff key
-		labels[common.LabelAnnotationVendor] = "zot-vendor"
-		labels[common.LabelAnnotationDescription] = "zot-label-desc"
+		labels[convert.LabelAnnotationVendor] = "zot-vendor"
+		labels[convert.LabelAnnotationDescription] = "zot-label-desc"
 		labels[ispec.AnnotationLicenses] = "zot-label-license"
 
-		desc = common.GetDescription(labels)
+		desc = convert.GetDescription(labels)
 		So(desc, ShouldEqual, "zot-label-desc")
 
-		license = common.GetLicenses(labels)
+		license = convert.GetLicenses(labels)
 		So(license, ShouldEqual, "zot-label-license")
 
-		vendor = common.GetVendor(labels)
+		vendor = convert.GetVendor(labels)
 		So(vendor, ShouldEqual, "zot-vendor")
 
-		routePrefix := common.GetRoutePrefix("test:latest")
+		routePrefix := storage.GetRoutePrefix("test:latest")
 		So(routePrefix, ShouldEqual, "/")
 
-		routePrefix = common.GetRoutePrefix("a/test:latest")
+		routePrefix = storage.GetRoutePrefix("a/test:latest")
 		So(routePrefix, ShouldEqual, "/a")
 
-		routePrefix = common.GetRoutePrefix("a/b/test:latest")
+		routePrefix = storage.GetRoutePrefix("a/b/test:latest")
 		So(routePrefix, ShouldEqual, "/a")
 
 		allTags, vulnerableTags := getTags()
 
-		latestTag := common.GetLatestTag(allTags)
+		latestTag := GetLatestTag(allTags)
 		So(latestTag.Name, ShouldEqual, "1.0.3")
 
-		fixedTags := common.GetFixedTags(allTags, vulnerableTags)
+		fixedTags := cveinfo.GetFixedTags(allTags, vulnerableTags)
 		So(len(fixedTags), ShouldEqual, 2)
 
-		fixedTags = common.GetFixedTags(allTags, append(vulnerableTags, common.TagInfo{
+		fixedTags = cveinfo.GetFixedTags(allTags, append(vulnerableTags, cvemodel.TagInfo{
 			Name:       "taginfo",
-			Descriptor: common.Descriptor{},
+			Descriptor: cvemodel.Descriptor{},
 			Timestamp:  time.Date(2000, time.July, 20, 10, 10, 10, 10, time.UTC),
 		}))
 		So(len(fixedTags), ShouldEqual, 3)
 
 		log := log.NewLogger("debug", "")
-
-		rootDir := t.TempDir()
 
 		subRootDir := t.TempDir()
 
@@ -1900,8 +1881,6 @@ func TestUtilsMethod(t *testing.T) {
 		conf.Extensions.Lint = &extconf.LintConfig{}
 
 		metrics := monitoring.NewMetricsServer(false, log)
-		defaultStore := local.NewImageStore(rootDir, false,
-			storage.DefaultGCDelay, false, false, log, metrics, nil, nil)
 
 		subStore := local.NewImageStore(subRootDir, false,
 			storage.DefaultGCDelay, false, false, log, metrics, nil, nil)
@@ -1909,16 +1888,6 @@ func TestUtilsMethod(t *testing.T) {
 		subStoreMap := make(map[string]storage.ImageStore)
 
 		subStoreMap["/b"] = subStore
-
-		storeController := storage.StoreController{DefaultStore: defaultStore, SubStore: subStoreMap}
-
-		dir := common.GetRootDir("a/zot-cve-test", storeController)
-
-		So(dir, ShouldEqual, rootDir)
-
-		dir = common.GetRootDir("b/zot-cve-test", storeController)
-
-		So(dir, ShouldEqual, subRootDir)
 
 		repo, digest := common.GetImageDirAndDigest("image")
 		So(repo, ShouldResemble, "image")
@@ -2390,7 +2359,7 @@ func TestGetImageManifest(t *testing.T) {
 		storeController := storage.StoreController{
 			DefaultStore: mockImageStore,
 		}
-		olu := NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
+		olu := ocilayout.NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
 
 		_, _, err := olu.GetImageManifest("nonexistent-repo", "latest")
 		So(err, ShouldNotBeNil)
@@ -2406,7 +2375,7 @@ func TestGetImageManifest(t *testing.T) {
 		storeController := storage.StoreController{
 			DefaultStore: mockImageStore,
 		}
-		olu := NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
+		olu := ocilayout.NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
 
 		_, _, err := olu.GetImageManifest("test-repo", "latest") //nolint:goconst
 		So(err, ShouldNotBeNil)
@@ -3068,7 +3037,7 @@ func TestGetRepositories(t *testing.T) {
 			DefaultStore: mockImageStore,
 			SubStore:     map[string]storage.ImageStore{"test": mockImageStore},
 		}
-		olu := NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
+		olu := ocilayout.NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
 
 		repoList, err := olu.GetRepositories()
 		So(repoList, ShouldBeEmpty)
@@ -3078,7 +3047,7 @@ func TestGetRepositories(t *testing.T) {
 			DefaultStore: mocks.MockedImageStore{},
 			SubStore:     map[string]storage.ImageStore{"test": mockImageStore},
 		}
-		olu = NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
+		olu = ocilayout.NewBaseOciLayoutUtils(storeController, log.NewLogger("debug", ""))
 
 		repoList, err = olu.GetRepositories()
 		So(repoList, ShouldBeEmpty)
@@ -3388,7 +3357,7 @@ func TestGlobalSearch(t *testing.T) {
 		)
 		So(err, ShouldBeNil)
 
-		olu := NewBaseOciLayoutUtils(ctlr.StoreController, log.NewLogger("debug", ""))
+		olu := ocilayout.NewBaseOciLayoutUtils(ctlr.StoreController, log.NewLogger("debug", ""))
 
 		// Initialize the objects containing the expected data
 		repos, err := olu.GetRepositories()
@@ -3717,7 +3686,7 @@ func TestGlobalSearch(t *testing.T) {
 		)
 		So(err, ShouldBeNil)
 
-		olu := NewBaseOciLayoutUtils(ctlr.StoreController, log.NewLogger("debug", ""))
+		olu := ocilayout.NewBaseOciLayoutUtils(ctlr.StoreController, log.NewLogger("debug", ""))
 
 		// Initialize the objects containing the expected data
 		repos, err := olu.GetRepositories()
